@@ -2,18 +2,6 @@
   "name": "project-flashcards",
   "version": "0.1.0",
   "private": true,
-  "dependencies": {
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.5.0",
-    "@testing-library/user-event": "^7.2.1",
-    "bootstrap-icons": "^1.3.0",
-    "cross-fetch": "^3.0.5",
-    "react": "^16.13.1",
-    "react-bootstrap-icons": "^1.3.0",
-    "react-dom": "^16.13.1",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "3.4.3"
-  },
   "scripts": {
     "start": "concurrently 'npm run start:server' 'npm run start:react'",
     "start:react": "react-scripts start",
@@ -37,8 +25,16 @@
       "last 1 safari version"
     ]
   },
-  "devDependencies": {
-    "concurrently": "^5.3.0",
-    "json-server": "^0.16.1"
+  "dependencies": {
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.3.0",
+    "@testing-library/user-event": "^14.4.3",
+    "bootstrap-icons": "^1.9.1",
+    "cross-fetch": "^3.1.5",
+    "react": "^18.2.0",
+    "react-bootstrap-icons": "^1.8.4",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "^5.0.1"
   }
 }
